More Linux 64 bit fixes.
git-svn-id: svn://0x263a.de/tekuum_src@227 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ game/gamesys/Class.cpp
@@ -988,48 +988,49 @@ http://developer.apple.com/documentation/DeveloperTools/Conceptual/MachORuntime/
 
 	assert( D_EVENT_MAXARGS == 8 );
 
+	// Techyon RB: 64 bit fixes, changed int to intptr_t
 	switch( ev->GetNumArgs() ) {
 	case 0 :
 		( this->*callback )();
 		break;
 
 	case 1 :
-		typedef void ( idClass::*eventCallback_1_t )( const int );
+		typedef void ( idClass::*eventCallback_1_t )( const intptr_t );
 		( this->*( eventCallback_1_t )callback )( data[ 0 ] );
 		break;
 
 	case 2 :
-		typedef void ( idClass::*eventCallback_2_t )( const int, const int );
+		typedef void ( idClass::*eventCallback_2_t )( const intptr_t, const intptr_t );
 		( this->*( eventCallback_2_t )callback )( data[ 0 ], data[ 1 ] );
 		break;
 
 	case 3 :
-		typedef void ( idClass::*eventCallback_3_t )( const int, const int, const int );
+		typedef void ( idClass::*eventCallback_3_t )( const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_3_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ] );
 		break;
 
 	case 4 :
-		typedef void ( idClass::*eventCallback_4_t )( const int, const int, const int, const int );
+		typedef void ( idClass::*eventCallback_4_t )( const intptr_t, const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_4_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ], data[ 3 ] );
 		break;
 
 	case 5 :
-		typedef void ( idClass::*eventCallback_5_t )( const int, const int, const int, const int, const int );
+		typedef void ( idClass::*eventCallback_5_t )( const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_5_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ], data[ 3 ], data[ 4 ] );
 		break;
 
 	case 6 :
-		typedef void ( idClass::*eventCallback_6_t )( const int, const int, const int, const int, const int, const int );
+		typedef void ( idClass::*eventCallback_6_t )( const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_6_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ], data[ 3 ], data[ 4 ], data[ 5 ] );
 		break;
 
 	case 7 :
-		typedef void ( idClass::*eventCallback_7_t )( const int, const int, const int, const int, const int, const int, const int );
+		typedef void ( idClass::*eventCallback_7_t )( const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_7_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ], data[ 3 ], data[ 4 ], data[ 5 ], data[ 6 ] );
 		break;
 
 	case 8 :
-		typedef void ( idClass::*eventCallback_8_t )( const int, const int, const int, const int, const int, const int, const int, const int );
+		typedef void ( idClass::*eventCallback_8_t )( const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t, const intptr_t );
 		( this->*( eventCallback_8_t )callback )( data[ 0 ], data[ 1 ], data[ 2 ], data[ 3 ], data[ 4 ], data[ 5 ], data[ 6 ], data[ 7 ] );
 		break;
 
@@ -1037,6 +1038,7 @@ http://developer.apple.com/documentation/DeveloperTools/Conceptual/MachORuntime/
 		gameLocal.Warning( "Invalid formatspec on event '%s'", ev->GetName() );
 		break;
 	}
+	// Techyon END
 
 #endif
